+++ src/skills/Entries.php
@@ -13,6 +13,7 @@ namespace doublesecretagency\sidekick\skills;
 
 use Craft;
 use craft\elements\Entry;
+use craft\helpers\Json;
 use doublesecretagency\sidekick\models\SkillResponse;
 use Throwable;
 
@@ -21,17 +22,31 @@ class Entries
     /**
      * Create a new entry.
      *
-     * @param string $title A creatively and randomly generated title for the entry.
+     * If you do not have a clear understanding of which sections exist, call the `getSections` skill first.
+     *
+     * The `$jsonConfig` parameter may contain any of the following data:
+     *
+     * {
+     *     "sectionId": 1, // Required
+     *     "typeId": 1, // Required
+     *     "title": "My New Entry", // Required
+     *     "slug": "my-new-entry",
+     *     "status": "live",
+     * }
+     *
      * @param string $jsonConfig A JSON string containing the configuration for the entry.
      * @return SkillResponse
      */
-    public static function createEntry(string $title, string $jsonConfig): SkillResponse
+    public static function createEntry(string $jsonConfig): SkillResponse
     {
+        // Decode the JSON configuration
+        $data = Json::decode($jsonConfig);
+
         // Configure the new entry
         $entry = new Entry();
-        $entry->sectionId = 1;
-        $entry->typeId = 1;
-        $entry->title = $title;
+        $entry->sectionId = ($data['sectionId'] ?? 1);
+        $entry->typeId = ($data['typeId'] ?? 1);
+        $entry->title = ($data['title'] ?? 'New Entry');
 
         // Optionally set additional field values:
         // $entry->setFieldValue('fieldHandle', 'Field Value');
@@ -56,7 +71,7 @@ class Entries
         // Return success message
         return new SkillResponse([
             'success' => true,
-            'message' => "Entry \"{$title}\" has been created. [{$jsonConfig}]",
+            'message' => "Entry \"{$entry->title}\" has been created. [{$jsonConfig}]",
 //            'response' => $config,
         ]);
     }