+++ src/skills/SettingsSections.php
@@ -36,10 +36,16 @@ class SettingsSections
     /**
      * Get a complete list of existing sections.
      *
+     * If you are unfamiliar with the existing sections, you MUST call this tool before creating, reading, updating, or deleting sections.
+     * Eagerly call this if an understanding of the current sections is required.
+     *
+     * You may also find it helpful to call this tool before updating an Entry.
+     *
      * @return SkillResponse
      */
     public static function getSections(): SkillResponse
     {
+        // Initialize sections
         $sections = [];
 
         // Fetch all sections
@@ -47,11 +53,30 @@ class SettingsSections
 
         // Loop through each section and format the output
         foreach ($allSections as $section) {
+
+            // Initialize entry types
+            $entryTypes = [];
+
+            // Get the entry types for the section
+            foreach ($section->getEntryTypes() as $entryType) {
+                // Catalog each entry type
+                $entryTypes[] = [
+                    'ID' => $entryType->id,
+                    'Name' => $entryType->name,
+                    'Handle' => $entryType->handle,
+                    'Field Layout' => $entryType->fieldLayoutId,
+                ];
+            }
+
+            // Catalog each section
             $sections[] = [
+                'ID' => $section->id,
                 'Name' => $section->name,
                 'Handle' => $section->handle,
-                'Section Type' => $section->type
+                'Section Type' => $section->type,
+                'Available Entry Types' => $entryTypes,
             ];
+
         }
 
         // Return success message